+++ src/main/kotlin/net/weavemc/loader/hooks/client/GuiOpenEventHook.kt
@@ -1,8 +1,8 @@
-package net.weavemc.loader.hooks
+package net.weavemc.loader.hooks.client
 
 import net.weavemc.loader.api.Hook
 import net.weavemc.loader.api.event.CancellableEvent
-import net.weavemc.loader.api.event.client.ChatEvent
+import net.weavemc.loader.api.event.client.GuiOpenEvent
 import net.weavemc.loader.api.util.asm
 import net.weavemc.loader.util.callEvent
 import net.weavemc.loader.util.internalNameOf
@@ -10,17 +10,17 @@ import net.weavemc.loader.util.named
 import org.objectweb.asm.tree.ClassNode
 import org.objectweb.asm.tree.LabelNode
 
-internal class ChatSentEventHook : Hook("net/minecraft/client/entity/EntityPlayerSP") {
+internal class GuiOpenEventHook : Hook("net/minecraft/client/Minecraft") {
     override fun transform(node: ClassNode, cfg: AssemblerConfig) {
-        node.methods.named("sendChatMessage").instructions.insert(asm {
-            new(internalNameOf<ChatEvent.Sent>())
+        node.methods.named("displayGuiScreen").instructions.insert(asm {
+            new(internalNameOf<GuiOpenEvent>())
             dup
             dup
             aload(1)
             invokespecial(
-                internalNameOf<ChatEvent.Sent>(),
+                internalNameOf<GuiOpenEvent>(),
                 "<init>",
-                "(L${internalNameOf<String>()};)V"
+                "(Lnet/minecraft/client/gui/GuiScreen;)V"
             )
             callEvent()
